+++ packages/react-docs/CHANGELOG.md
@@ -3,6 +3,137 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.16...@patternfly/react-docs@5.3.17) (2020-06-03)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.15...@patternfly/react-docs@5.3.16) (2020-06-01)
+
+
+### Bug Fixes
+
+* **breakpoints:** make API consistent and appear in docs ([#4310](https://github.com/patternfly/patternfly-react/issues/4310)) ([ac1afa6](https://github.com/patternfly/patternfly-react/commit/ac1afa60067346c658fcb160534d9ddaf7dafbc6))
+
+
+
+
+
+## [5.3.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.14...@patternfly/react-docs@5.3.15) (2020-05-29)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.13...@patternfly/react-docs@5.3.14) (2020-05-29)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.12...@patternfly/react-docs@5.3.13) (2020-05-28)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.11...@patternfly/react-docs@5.3.12) (2020-05-27)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.10...@patternfly/react-docs@5.3.11) (2020-05-27)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.9...@patternfly/react-docs@5.3.10) (2020-05-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.8...@patternfly/react-docs@5.3.9) (2020-05-21)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.7...@patternfly/react-docs@5.3.8) (2020-05-21)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.6...@patternfly/react-docs@5.3.7) (2020-05-21)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.5...@patternfly/react-docs@5.3.6) (2020-05-21)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.4...@patternfly/react-docs@5.3.5) (2020-05-21)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.3...@patternfly/react-docs@5.3.4) (2020-05-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.2...@patternfly/react-docs@5.3.3) (2020-05-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [5.3.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.1...@patternfly/react-docs@5.3.2) (2020-05-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [5.3.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@5.3.0...@patternfly/react-docs@5.3.1) (2020-05-18)