+++ packages/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,101 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.13.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.13.1...demo-app-ts@4.13.2) (2020-06-03)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.13.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.13.0...demo-app-ts@4.13.1) (2020-06-01)
+
+
+### Bug Fixes
+
+* **breakpoints:** make API consistent and appear in docs ([#4310](https://github.com/patternfly/patternfly-react/issues/4310)) ([ac1afa6](https://github.com/patternfly/patternfly-react/commit/ac1afa60067346c658fcb160534d9ddaf7dafbc6))
+
+
+
+
+
+# [4.13.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.12.4...demo-app-ts@4.13.0) (2020-05-28)
+
+
+### Features
+
+* **Form:** remove semicolons from docs ([#4296](https://github.com/patternfly/patternfly-react/issues/4296)) ([c666474](https://github.com/patternfly/patternfly-react/commit/c666474bad1741fbd30c3ceccaa3eafc5187bd72))
+
+
+
+
+
+## [4.12.4](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.12.3...demo-app-ts@4.12.4) (2020-05-27)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.12.3](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.12.2...demo-app-ts@4.12.3) (2020-05-26)
+
+
+### Bug Fixes
+
+* **wiz:** add missing label for wiz-in-modal ([#4280](https://github.com/patternfly/patternfly-react/issues/4280)) ([71a45d1](https://github.com/patternfly/patternfly-react/commit/71a45d122b20afddb9f2c99105f3603cbfc96bac))
+
+
+
+
+
+## [4.12.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.12.1...demo-app-ts@4.12.2) (2020-05-21)
+
+
+### Bug Fixes
+
+* **charts:** Use Victory's createContainer instead of allowZoom prop ([#4278](https://github.com/patternfly/patternfly-react/issues/4278)) ([8c353d9](https://github.com/patternfly/patternfly-react/commit/8c353d9a5a4322984d5e564d5737e53e796928f6))
+
+
+
+
+
+## [4.12.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.12.0...demo-app-ts@4.12.1) (2020-05-21)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+# [4.12.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.11.2...demo-app-ts@4.12.0) (2020-05-21)
+
+
+### Features
+
+* **drawer:** use the breakpoint mods prop instead of width props ([#4270](https://github.com/patternfly/patternfly-react/issues/4270)) ([4cac8d0](https://github.com/patternfly/patternfly-react/commit/4cac8d01ea771a7e8b082ca33242fcf590b861f8))
+
+
+
+
+
+## [4.11.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.11.1...demo-app-ts@4.11.2) (2020-05-21)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.11.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.11.0...demo-app-ts@4.11.1) (2020-05-19)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
 # [4.11.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@3.31.7...demo-app-ts@4.11.0) (2020-05-15)